+++ src/app/admin/contact-us/contact-us.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatTableModule } from '@angular/material/table';
+import { MatTooltipModule } from '@angular/material/tooltip';
 import { ContactUs } from 'src/app/models/contactUs';
 import { DataService } from 'src/app/service/data.service';
 import { environment } from 'src/environment/environment';
@@ -12,7 +13,7 @@ import { environment } from 'src/environment/environment';
   templateUrl: './contact-us.component.html',
   styleUrls: ['./contact-us.component.css'],
   standalone: true,
-  imports:[NgIf,MatIconModule,MatButtonModule,MatTableModule]
+  imports:[NgIf,MatIconModule,MatButtonModule,MatTableModule,MatTooltipModule]
 })
 export class ContactUsComponent implements OnInit {
   contact:ContactUs[]=[];
@@ -26,9 +27,7 @@ export class ContactUsComponent implements OnInit {
   ];
     constructor(private service:DataService){ }
   ngOnInit(): void {
-    this.service.GetJsonItem<ContactUs[]>(environment.contactUs.get).subscribe(data=>{
-      this.contact = data
-    })
+    this.getItem()
   }
   getItem(){
     this.service.GetJsonItem<ContactUs[]>(environment.contactUs.get).subscribe(data=>{
@@ -36,8 +35,10 @@ export class ContactUsComponent implements OnInit {
     })
   }
   del(id:number){
-    this.service.DeleteItem(`${environment.contactUs.get}/${id}`).subscribe(()=>{
-      this.getItem()
-    })
+   if (confirm('Do you really want to delete')) {
+     this.service.DeleteItem(`${environment.contactUs.get}/${id}`).subscribe(()=>{
+       this.getItem()
+     })
+   }
   }
 }
